# Get the grains that we check on every startup from the start event