--- a/java/spacewalk-java.changes
+++ b/java/spacewalk-java.changes
@@ -1,3 +1,4 @@
+- Get the machineid grain from the minion startup event
 - Feat: enable Salt system lock when CaaSP node is onboarded
 - use term 'patch' instead of 'errata' (bsc#1164649)
 - enable provisioning API with salt and bootstrap entitled systems