--- a/susemanager-utils/susemanager-sls/salt/bootstrap/susemanager.conf
+++ b/susemanager-utils/susemanager-sls/salt/bootstrap/susemanager.conf
@@ -8,3 +8,7 @@ grains:
   susemanager:
     activation_key: {{ pillar['activation_key'] }}
 {% endif %}
+start_event_grains:
+  - machine_id
+  - saltboot_initrd
+  - susemanager