--- a/susemanager-utils/susemanager-sls/salt/util/mgr_start_event_grains.sls
+++ b/susemanager-utils/susemanager-sls/salt/util/mgr_start_event_grains.sls
@@ -1,0 +1,6 @@
+mgr_start_event_grains:
+  file.append:
+    - name: /etc/salt/minion.d/susemanager.conf
+    - text: |
+        start_event_grains:
+          - machine_id