--- a/susemanager-utils/susemanager-sls/salt/util/mgr_start_event_grains.sls
+++ b/susemanager-utils/susemanager-sls/salt/util/mgr_start_event_grains.sls
@@ -4,3 +4,5 @@ mgr_start_event_grains:
     - text: |
         start_event_grains:
           - machine_id
+          - saltboot_initrd
+          - susemanager