… java/code/src/com/suse/manager/reactor/test/RegisterMinionActionTest.java
… java/code/src/com/suse/manager/webui/utils/salt/MinionStartEvent.java
… java/code/src/com/suse/manager/webui/utils/salt/MinionStartupGrains.java
… spacewalk/certs-tools/spacewalk-certs-tools.changes