## What does this PR change?
Get the grains (`machine_id`, `saltboot_initrd`, `management_key`) that we check on every startup from the start event instead of making a separate call. 

## Background
We faced performance issues during minions startup as `machine_id` grains is needed to lookup if minion already exists in the database. This is needed to differentiate between a startup and 1st-time registration.

We followed different approaches, each with some pros/cons. Those approaches are described herein this PR description https://github.com/SUSE/spacewalk/pull/8892. The related research card is https://github.com/SUSE/spacewalk/issues/8228m where one can see further discussion.

In the end, we introduced a patch in upstream which makes it possible to pass interesting grains in the payload of minion start-up event. Related PR https://github.com/saltstack/salt/pull/54948.
We think this is better approach in a sense that, we don't need to make a separate call, of course, it has a downside that all the minions need to have this new parameter in their config so next time, when minion get started, we use grains from the start_up event instead of making a separate call.

It is also noticeable that I kept the old way as a fallback, the reason was the older minions might not get the new changes so we have to have a backup plan.

We save 1 salt call in case of already registered minions, even if startup grains are not configured.

So if start-up grains are **supported**, with this PR number of salt sync calls are following


Test                                                |Before |After 
|-----------------------------------------------------|----------|----------|
| New registration                     | 2 | 1 |
| Registered minion startup                        | 2 | 0 |

If start-up grains are **not supported**, with this PR number of salt sync calls are following

Test                                                |Before |After 
|-----------------------------------------------------|----------|----------|
| New registration                     | 2 | 2 |
| Registered minion startup                        | 2 |  1 |


### Benchmarks

According to tests with 100 evil-minions at normal speed mode (`--slowdown-factor 1`) an **~1.32x speedup** on the total time to handle all minion start events (29s on average versus 38.5s).

Note that it is expected that the slower the simulated minions, the bigger the speedup.

## GUI diff

No difference.

- [x] **DONE**

## Documentation
-**Instead of a separate call, the grain is part of startup event, there is nothing to document here for the customer. I will add the documentation for using the helper `mgr_start_event_grains.sls state though**
 https://github.com/uyuni-project/uyuni-docs/pull/88

- [x] **DONE**

## Test coverage
- No tests: **Existing tests are modified**

- [x] **DONE**

## Links

Fixes # https://github.com/SUSE/spacewalk/issues/9909
Tracks # **add downstream PR, if any**

- [ ] **DONE**