@admd 2020-02-04
Thank you @moio for the review. I really overlooked some part. I have made the suggested changed. Please have a  look. Now the only remaining part is the PR for sumaform & for doc. I will create them in a while.