@admd 2020-02-05
Thank you @moio for the review. I have added 2 more commits with the changes. You can review them separately.