@admd 2020-02-21
@mcalmer I have rebased the branch & you can review it now. There is some untreated rubocop error but rest should be fine.